Switch from ssh to https for all git repos
While we directed the docker image to translate the URLs to https anyway
this can be confusing since git behaves differently in docker vs outside.

Also in the meantime we have made all repos public, so there is no auth needed
for the repos and the extra work required using https no longer applies.