docker: remove redis/redis-commander
In all our installations we ended up just using mariadb for the queue
and locking, and performance has been good enough.

To avoid confusion if redis is required or not and to reduce the startup
time of the docker-compose project just remove the redis services.

Users can add it back in their fork if needed.